refactor(core): group the install lifecycle into core/install
Move install-source detection, per-channel release lookup, and the
hunk update execution out of core/process into a core/install module:
they are one feature family about how the binary was installed and how
it gets replaced, not about the process a run lives in, and they only
landed in process because the startup update notice lived there. The
notice stays in core/process as a startup-notice producer built on the
app-state file and consumes core/install. Path-only move; no exported
symbol changed.
