+++ src/core/install/installSource.ts
