+++ src/core/install/latestRelease.test.ts
