+++ src/core/install/latestRelease.ts
