+++ src/core/install/selfUpdate.test.ts
