+++ src/core/install/selfUpdate.ts
